+++ system/src/Grav/Common/Debugger.php
@@ -762,7 +762,7 @@ class Debugger
 
             // Only render the bar if the page is HTML.
             $page = $this->grav['page'];
-            if ($page->templateFormat() !== 'html') {
+            if (!$page instanceof PageInterface || $page->templateFormat() !== 'html') {
                 return $this;
             }